+++ docs/.vuepress/config.js
@@ -1,6 +1,7 @@
 module.exports = {
   title: "vue-composable",
-  description: "Vue composition-api composable components",
+  description:
+    "Vue composition-api composable components. i18n, validation, pagination, fetch, etc. +30 variable composable functions",
   plugins: [
     "@vuepress/back-to-top",
     [
@@ -11,7 +12,7 @@ module.exports = {
       }
     ]
   ],
-  base: "/vue-composable/",
+  base: process.env.DOCS_BASE || "",
   head: [
     ["meta", { charset: "utf-8" }],
     ["meta", { name: "msapplication-TileColor", content: "#ffffff" }],
@@ -103,7 +104,10 @@ module.exports = {
           title: "Format",
           sidebarDepth: 1,
           collapsable: false,
-          children: [["composable/format/format", "format"]]
+          children: [
+            ["composable/format/format", "format"],
+            ["composable/format/path", "path"]
+          ]
         },
 
         {
@@ -149,8 +153,18 @@ module.exports = {
           collapsable: false,
           children: [
             ["composable/promise/promise", "promise"],
+            ["composable/promise/promiseLazy", "Lazy Promise"],
             ["composable/promise/cancellablePromise", "cancellablePromise"],
             ["composable/promise/retry", "retry"]
+          ]
+        },
+        {
+          title: "Intl",
+          sidebarDepth: 1,
+          collapsable: false,
+          children: [
+            ["composable/Intl/numberFormat", "numberFormat"],
+            ["composable/Intl/currencyFormat", "currencyFormat"]
           ]
         },
         {
@@ -166,7 +180,10 @@ module.exports = {
             ["composable/web/pageVisibility", "PageVisibilityAPI"],
             ["composable/web/language", "Language"],
             ["composable/web/broadcastChannel", "BroadcastChannel API"],
-            ["composable/web/geolocation", "Geolocation API"]
+            ["composable/web/geolocation", "Geolocation API"],
+            ["composable/web/cssVariables", "CSS variables"],
+            ["composable/web/worker", "WebWorker API"],
+            ["composable/web/workerFunction", "WebWorker Function"]
           ]
         },
         {