+++ examples/vue-composable-example/src/components/Validation.vue
@@ -15,12 +15,12 @@
 </template>
 
 <script>
-import { createComponent, ref, reactive } from "@vue/composition-api";
+import { defineComponent, ref, reactive } from "@vue/composition-api";
 import { useValidation } from "vue-composable";
 
 const required = x => !!x;
 
-export default createComponent({
+export default defineComponent({
   setup() {
     const name = ref("pikax");
     const surname = ref("stuff");