+++ packages/core/__tests__/promise/promise.spec.ts
@@ -1,5 +1,6 @@
 import { usePromise } from "../../src/promise/promise";
 import { nextTick } from "../utils";
+import { promisedTimeout } from "../../src/";
 
 describe("promise", () => {
   let cb: ((...args: any[]) => void) | null = null;
@@ -11,8 +12,26 @@ describe("promise", () => {
     currentPromise = null;
   });
 
-  it("should not resolve immediately", () => {
+  it("should resolve immediately", async () => {
     expect(usePromise(factory)).toMatchObject({
+      promise: { value: currentPromise },
+      result: { value: null },
+      loading: { value: true },
+      error: { value: null }
+    });
+  });
+
+  it("should not resolve immediately", () => {
+    expect(usePromise(factory, true)).toMatchObject({
+      promise: { value: undefined },
+      result: { value: null },
+      loading: { value: false },
+      error: { value: null }
+    });
+  });
+
+  it("should not resolve immediately with options", () => {
+    expect(usePromise(factory, { lazy: true })).toMatchObject({
       promise: { value: undefined },
       result: { value: null },
       loading: { value: false },
@@ -73,7 +92,7 @@ describe("promise", () => {
       error: { value: null }
     });
 
-    callbacks[1](1);
+    callbacks[2](1);
     await nextTick();
 
     expect(use).toMatchObject({
@@ -154,7 +173,7 @@ describe("promise", () => {
 
     const arr = [use.exec(), use.exec()];
 
-    callbacks[0](1);
+    callbacks[1](1);
     expect(await arr[0]).toBe(1);
   });
 
@@ -182,12 +201,54 @@ describe("promise", () => {
     });
   });
 
+  it("should only update result after the promise is resolved", async () => {
+    // result should not be set to `null` between executions
+    let v = 1;
+    const use = usePromise(x => promisedTimeout(20).then(() => (v = x)));
+
+    expect(use.result.value).toBe(null);
+
+    await use.exec(12);
+
+    expect(use.result.value).toBe(v);
+    use.exec(1);
+    await promisedTimeout(5);
+    expect(use.result.value).toBe(12);
+    await promisedTimeout(17);
+    expect(use.result.value).toBe(1);
+  });
+
+  describe("__DEV__ warns", () => {
+    const consoleWarnSpy = jest.spyOn(console, "warn");
+
+    it("should warn if the factory has argument but lazy no specified", () => {
+      usePromise((a: any) => {});
+      expect(consoleWarnSpy).toHaveBeenCalledWith(
+        "[usePromise] parameters detected on `fn` factory. Executing promise without arguments."
+      );
+
+      consoleWarnSpy.mockClear();
+      usePromise((a: any) => {}, true);
+      expect(consoleWarnSpy).not.toHaveBeenCalled();
+
+      usePromise((a: any) => {}, { lazy: true });
+      expect(consoleWarnSpy).not.toHaveBeenCalled();
+
+      usePromise((a: any) => {}, { throwException: true });
+      expect(consoleWarnSpy).toHaveBeenCalledWith(
+        "[usePromise] parameters detected on `fn` factory. Executing promise without arguments."
+      );
+    });
+  });
+
   describe("throw exception", () => {
     it("should throw when throwException is true at creation", async () => {
       expect.assertions(2);
 
       const error = new Error("error");
-      const use = usePromise(() => Promise.reject(error), true);
+      const use = usePromise(() => Promise.reject(error), {
+        throwException: true
+      });
 
       try {
         await use.exec();