+++ packages/core/src/format/format.ts
@@ -10,6 +10,11 @@ export interface FormatObject {
 }
 
 export function useFormat(
+  format: RefTyped<Readonly<string>>,
+  obj?: RefTyped<FormatObject>
+): Readonly<Ref<string>>;
+
+export function useFormat(
   format: Readonly<RefTyped<string>>,
   obj?: RefTyped<FormatObject>
 ): Readonly<Ref<string>>;
@@ -22,6 +27,11 @@ export function useFormat(
 export function useFormat(
   format: Readonly<RefTyped<string>>,
   obj?: RefTyped<FormatObject> | Array<FormatValue>
+): Readonly<Ref<string>>;
+
+export function useFormat(
+  format: RefTyped<string>,
+  args: any
 ): Readonly<Ref<string>>;
 
 export function useFormat(