+++ packages/core/src/i18n/i18n.ts
@@ -17,6 +17,7 @@ import {
 } from "../utils";
 import { usePath, useFormat, FormatObject, FormatValue } from "../format";
 
+// istanbul ignore next
 const I18n_ACCESS_SYMBOL: InjectionKey<i18nResult<
   string[],
   string
@@ -52,7 +53,9 @@ interface i18nDefinition<TMessage> {
 
   fallback?: keyof TMessage;
 
-  messages: { [K in keyof TMessage]: i18n | (() => Promise<i18n>) };
+  messages: {
+    [K in keyof TMessage]: i18n | (() => Promise<i18n>) | (() => i18n);
+  };
 
   /**
    * Resolves the translation for i18n
@@ -78,15 +81,23 @@ interface i18nResult<TLocales, TMessages extends any = i18n> {
 
   $t(path: string, args?: object | Array<object>): Readonly<Ref<string>>;
 
+  $ts(path: string, args?: object | Array<object>): string;
+
   addLocale(locale: string, messages: TMessages): void;
   removeLocale(locale: TLocales): void;
 }
+
+type PromiseResult<T> = T extends Promise<infer R> ? R : T;
+
+type I18nExtractLocale<T> = T extends (...args: any[]) => any
+  ? PromiseResult<ReturnType<T>>
+  : PromiseResult<T>;
 
 export function useI18n<
   T extends i18nDefinition<TMessage>,
   TMessage extends Record<keyof T["messages"], i18n | (() => Promise<any>)>
 >(definition: T): i18nResult<keyof T["messages"], T["messages"][T["locale"]]>;
-export function useI18n<T = i18n>(): i18nResult<string[], T> | void;
+export function useI18n<T = i18n>(): i18nResult<string[], T>;
 export function useI18n(definition?: any): any {
   if (definition) {
     return buildI18n(definition);
@@ -95,42 +106,49 @@ export function useI18n(definition?: any): any {
 
 export function buildI18n<
   T extends i18nDefinition<TMessage>,
-  TMessage extends Record<keyof T["messages"], i18n | (() => Promise<any>)>,
-  TLocales extends keyof TMessage = keyof TMessage
->(definition: T): i18nResult<keyof T["messages"], T["messages"][T["locale"]]> {
-  const locales: Ref<Array<TLocales>> = ref(
-    Object.keys(definition.messages) as Array<TLocales>
+  TMessage extends Record<keyof T["messages"], i18n | (() => Promise<any>)>
+>(
+  definition: T
+): i18nResult<
+  keyof T["messages"],
+  I18nExtractLocale<T["messages"][T["locale"]]>
+> {
+  const locales = ref<Array<keyof TMessage>>(
+    Object.keys(definition.messages) as any
   );
-  const localeMessages: Ref<TMessage> = ref(definition.messages) as any;
-  const locale: Ref<TLocales> = ref<TLocales>(
-    definition.locale as TLocales
-  ) as any;
+  const localeMessages = ref<
+    Record<keyof TMessage, i18n | (() => Promise<any>)>
+  >(definition.messages as any);
+  const locale: Ref<keyof TMessage> = ref(definition.locale as any);
   const i18n = ref<any>({});
   let fallback = ref<i18n>();
 
-  const cache: Partial<Record<TLocales, Ref<i18n>>> = {};
+  const cache: Record<string, Ref<i18n>> = {};
 
   const loadLocale = (
-    locale: TLocales,
-    localeMessages: Ref<TMessage>
+    locale: string,
+    messages: typeof localeMessages
   ): Ref<i18n> | Promise<Ref<i18n>> => {
     if (cache[locale]) {
-      return cache[locale]!;
+      return cache[locale];
     }
 
-    const l = localeMessages.value[locale];
+    const l = messages.value[locale];
     if (!l) {
       return ref({});
     }
 
-    if (isFunction(l)) {
-      return Promise.resolve((l as Function)()).then(
-        x => (cache[locale] = wrap<i18n>(x))
-      );
+    let m = isFunction(l) ? l() : l;
+    if (isPromise(m)) {
+      return m.then(x => (cache[locale] = wrap<i18n>(x)));
     }
-    return (cache[locale] = computed(
-      () => localeMessages.value[locale] as i18n
-    ));
+
+    // if it was function we don't keep track on that
+    if (isFunction(l)) {
+      return wrap(m);
+    }
+
+    return (cache[locale] = computed(() => messages.value[locale]));
   };
 
   const shouldFallback = definition.fallback
@@ -141,26 +159,35 @@ export function buildI18n<
 
   let fallbackIsPromise = false;
   if (shouldFallback) {
-    const fallbackI18n = loadLocale(locale.value, localeMessages);
+    const fallbackI18n = loadLocale(
+      definition.fallback! as string,
+      localeMessages
+    );
     if (isPromise(fallbackI18n)) {
       fallbackI18n.then(x => {
-        fallback = x;
+        fallback.value = x.value;
       });
       fallbackIsPromise = true;
     } else {
-      fallback = fallbackI18n;
+      fallback.value = fallbackI18n.value;
     }
   } else {
     fallback.value = {};
   }
 
+  const localeChangesCount = ref(0);
+  watch(localeMessages, () => localeChangesCount.value++, {
+    deep: true,
+    immediate: false
+  });
+
   watch(
-    [locale, fallback] as const,
-    async ([l, fb]) => {
+    [locale, fallback, localeChangesCount],
+    async ([l, fb, _]: [keyof TMessage, i18n | undefined, any]) => {
       if (l === definition.fallback && shouldFallback) {
         i18n.value = fb!;
       } else {
-        const localeMessage = await loadLocale(l, localeMessages);
+        const localeMessage = await loadLocale(l as string, localeMessages);
         i18n.value = deepClone<any>({}, fb, localeMessage.value);
       }
     },
@@ -176,32 +203,82 @@ export function buildI18n<
     if (definition.resolve) {
       return wrap(definition.resolve(i18n.value, path, args));
     }
-    return useFormat(usePath(i18n, path, ".", (_, _1, p) => p) as any, args);
+    return useFormat(
+      usePath(i18n, path, ".", (_, _1, p, _2) => p) as Ref<any>,
+      args
+    );
+  };
+
+  const $ts = (
+    path: Readonly<RefTyped<string>>,
+    args: RefTyped<FormatObject> | Array<FormatValue> | undefined
+  ): string => {
+    return $t(path, args).value;
   };
 
   const addLocale = (l: string, m: any) => {
-    if (locales.value.indexOf(l as TLocales) >= 0) {
+    if (locales.value.indexOf(l as any) >= 0) {
+      /* istanbul ignore else */
       if (__DEV__) {
-        console.warn("Locale already exists, overriding it");
+        console.warn(
+          `[useI18n] Locale "${l}" already exists, overriding it...`
+        );
       }
     } else {
       locales.value.push(l as any);
     }
-    delete cache[l as TLocales];
-    (localeMessages.value as Record<string, i18n>)[l] = m;
+    delete cache[l];
+
+    localeMessages.value = {
+      ...localeMessages.value,
+      [l]: m
+    };
+    // Vue.set(localeMessages.value, l, m);
+    // (localeMessages.value as Record<string, i18n>)[l] = m;
   };
 
-  const removeLocale = (l: TLocales) => {
+  const removeLocale = (l: keyof TMessage) => {
     const index = locales.value.indexOf(l);
     if (index >= 0) {
+      const nextLocale = [
+        locale.value,
+        fallback.value && definition.fallback,
+        ...locales.value
+      ].find(x => x && x !== l);
+
+      if (nextLocale) {
+        if (l === definition.fallback) {
+          /* istanbul ignore else */
+          if (__DEV__) {
+            console.warn(`[useI18n] removing default fallback locale "${l}"`);
+          }
+          fallback.value = undefined;
+        }
+        if (l === locale.value) {
+          /* istanbul ignore else */
+          if (__DEV__) {
+            console.warn(
+              `[useI18n] removing current locale "${l}", setting current locale to "${nextLocale}"`
+            );
+          }
+          locale.value = nextLocale;
+        }
+      } else {
+        /* istanbul ignore else */
+        if (__DEV__) {
+          console.error("[useI18n] No locales available to use");
+        }
+      }
       locales.value.splice(index, 1);
     } else {
+      /* istanbul ignore else */
       if (__DEV__) {
-        console.warn("Locale doesn't exist");
+        console.warn(`[useI18n] Locale "${l}" doesn't exist`);
       }
     }
+    // Vue.delete(localeMessages.value, l as string);
     delete localeMessages.value[l];
-    delete cache[l];
+    delete cache[l as string];
   };
 
   return {
@@ -211,6 +288,7 @@ export function buildI18n<
     i18n,
 
     $t,
+    $ts,
 
     addLocale,
     removeLocale