+++ packages/web/src/storage/storage.ts
@@ -11,18 +11,19 @@ let canUseLocalStorage: boolean | undefined = undefined;
 
 export function useStorage(
   key: string,
-  defaultValue?: RefTyped<string>
+  defaultValue?: RefTyped<string>,
+  sync?: boolean
 ): LocalStorageReturn<string>;
 export function useStorage<T extends object = any>(
   key: LocalStorageTyped<T> | string,
-  defaultValue?: RefTyped<T>
+  defaultValue?: RefTyped<T>,
+  sync?: boolean
 ): LocalStorageReturn<T>;
-export function useStorage(key: string, defaultValue?: any) {
+export function useStorage(key: string, defaultValue?: any, sync?: boolean) {
   if (canUseLocalStorage === undefined) {
     canUseLocalStorage = useWebStorage("localStorage").supported;
   }
-
   return canUseLocalStorage
-    ? useLocalStorage(key, defaultValue)
-    : useSessionStorage(key, defaultValue);
+    ? useLocalStorage(key, defaultValue, sync)
+    : useSessionStorage(key, defaultValue, sync);
 }